wire doc description to docs plugin UI
