+++ packages/docusaurus-theme-classic/src/theme/BlogTagsPostsPage/index.tsx
@@ -57,7 +57,7 @@ function BlogTagsPostsPageMetadata({tag}: Props): JSX.Element {
   const title = useBlogTagsPostsPageTitle(tag);
   return (
     <>
-      <PageMetadata title={title} />
+      <PageMetadata title={title} description={tag.description} />
       <SearchMetadata tag="blog_tags_posts" />
     </>
   );
