+++ website/_dogfooding/_docs tests/tags.yml
@@ -1,4 +1,5 @@
 a:
+  description: 'Description for tag a'
 b:
 c:
 e:
